One more DBCStores.cpp fix, Global fixing warnings (signed/unsigned comparison, unused vars, changed field initialization in constructor).
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1698 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/shared/Auth/AuthCrypt.h
@@ -28,8 +28,8 @@ class AuthCrypt
         AuthCrypt();
         ~AuthCrypt();
 
-        const static int CRYPTED_SEND_LEN = 4;
-        const static int CRYPTED_RECV_LEN = 6;
+        const static size_t CRYPTED_SEND_LEN = 4;
+        const static size_t CRYPTED_RECV_LEN = 6;
 
         void Init();
